+++ ecommerce-inventory-service-api/src/main/resources/db/migration/V1.0__init.sql
