+++ ecommerce-inventory-service-sdk/src/main/java/com/ecommerce/inventory/sdk/representation/about/AboutRepresentation.java
@@ -1,0 +1,15 @@
+package com.ecommerce.inventory.sdk.representation.about;
+
+
+import lombok.Value;
+
+@Value
+public class AboutRepresentation {
+    private String buildNumber;
+    private String buildTime;
+    private String deployTime;
+    private String gitRevision;
+    private String gitBranch;
+    private String environment;
+
+}
